pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "juggler"
version = "0.2.0"
description = "🤹 Juggle multiple AI providers like a pro - Chat, embeddings, vision, and more"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Spandan Bhol", email = "spandankb@gmail.com"}
]
keywords = ["ai", "llm", "routing", "groq", "cerebras", "mistral", "nvidia", "cohere", "embeddings", "reranking", "transcription", "tts", "multi-provider","anyllm","openrouter","litellm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "responses>=0.23.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools]
packages = ["juggler"]

[project.urls]
Homepage = "https://github.com/spanDevOps/juggler"
Documentation = "https://github.com/spanDevOps/juggler#readme"
Repository = "https://github.com/spanDevOps/juggler"
Issues = "https://github.com/spanDevOps/juggler/issues"

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
